Extract manifest schemas into dedicated module and add v0.4 support
Move all manifest-related Zod schemas (ServerType, UserConfigField,
McpConfig, ManifestAuthor, ManifestServer, Capability, McpbManifest)
out of package.ts into a new manifest.ts as the single source of truth.
package.ts re-exports these for backward compatibility.

The upstream MCPB spec v0.3 and v0.4 are backward-compatible (v0.4 only
adds "uv" as a server runtime type), so a single permissive schema
handles both versions without a discriminated union.

- Add "uv" to ServerTypeSchema enum
- Add "uv" case to SDK resolveCommand (defaults to `uv run <entry_point>`)
- Export manifest.ts from schemas barrel
